ruleset package moved into the provider package
